Fix rule circular reference memory leak.
If you specify a state with a rule that point to a state that also has a rule
that eventually point back to the first state, well, you'll get a memory
leak. The fix is to weaken references to states in rules.

Many thanks to @glasswalk3r for finally chasing down the proper fix, which in
turn allowed me to do some poking to reproduce the leak as a test failure so
that it can stay fixed.

Closes #6.