Explain how we test for leaks.
It takes advantage of the internals, using a string to mimic a memory address.
But since it's those internals that cause the leak in the first place, I think
this is a reaonable way to do the test. However, it could be confusing without
signifant study, so leave a note for my future self to explain how this works
and why.